requirements.txt
PyMuPDF
Pillow
moviepy
numpy
imageio
imageio-ffmpeg